ZRLE: implement run-length encoding of zero bytes to shrink patches
The bsdiff algorithm produces patches with _lots_ of zero bytes.
Tarballs also tend to have long runs of zeros in their data, so
it's common that the new data will include lots of zeros. Indeed,
running ZRLE alone on our reference patch reduces it by 99%, and
makes compressing what's left _much_ faster. Some compression
algorithms are almost as good as ZRLE + compression, but most are
not, so ZRLE is a highly effective preprocessing step.