rename: bssort => bsindex with header (#16)
This renames `bssort` to the more generic `bsindex` on the premise that we may, in the future, want to save more data in an index file. To future-proof it, this also adds a header to the file so that we can know if the file we're reading has the right format or not.