API: eliminate `*_core` methods by using `nothing` for optional args
This also refactors the format autodetection code to avoid the need
to ever rewind the patch I/O stream by looking at incrementally more
of the patch data until it either matches some magic string or can't
possibly match any of them.

Adds an (internal) API for format registration. If someone does want
to implement a new patch format and have it hook into this package,
this allows that to be done cleanly.