JUILA_BSDIFF_LOWMEM: work around low memory failures (#40)
This PR adds support for using the environment variable
JULIA_BSDIFF_LOWMEM to cause BSDiff to use less memory for both
generating and applying diffs. Ideally this shouldn't be necessary,
but it seems that libbzip2 will crash in low memory systems when
called repeatedly. I have tried hooking libbzip2's memory allocation
into Julia's allocator which should allow it to share memory that the
GC has allocated, as well as fixing various allocation issue in
CodecBzip2 but that doesn't fix the issue, hence this workaround. It's
unclear if this is a bug in libbzip2 on in the way that CodecBzip2
calls it. If, at some point, the memory bug in one of those is fixed,
we could remove this workaround.